:truck:rename: 컴포넌트 폴더명 오타 수정
