+++ components/board/Article.tsx
