+++ pages/_app.tsx
@@ -1,4 +1,4 @@
-import Header from "@/componets/common/Header";
+import Header from "@/components/common/Header";
 import "@/styles/globals.css";
 import "@/styles/reset.css";
 import type { AppProps } from "next/app";
