+++ styles/SearchForm.module.scss
@@ -6,7 +6,7 @@
   border: none;
   background: var(--cool-gray100)
     url(../public/assets/icons/item-search-icon.svg) no-repeat left 16px center;
-
+  outline: none;
   &::placeholder {
     font-weight: 400;
     font-size: 16px;
